+++ backend/templates/create-report/template-files/models/index.js
@@ -1,18 +1,33 @@
+// Libs
 const dbConnection = require("@database/databaseConnection");
 const { knex } = dbConnection();
 
-/**
- * Gets data for the $reportName report
- *
- * @param   {number[]} portfolios Optional list of portfolio_ids to limit report to. If empty, returns data for all portfolios.
- * @returns {any[]}
- */
+//Constants
+const name = `$reportName`;
 
-const $reportName = (portfolios) => {
-  const query = knex.raw();
-  return query;
+// gets the current date in ISO "YYYY-MM-DD" format.
+const getCurrentDate = async () => new Date().toISOString().split("T")[0];
+
+/**
+ * Retrieves the data for various financial metrics based on the fiscal year.
+ *
+ * @param   {number | string | Array} fiscal - The fiscal year(s) to retrieve totals for.
+ * @returns {Promise}                        - A promise that resolves to the query result containing the totals for recoveries, expenses, net recoveries, and quarterly gross and net amounts.
+ */
+const queries = {
+  [name]: (fiscal) => {
+    const query = knex.select(knex.raw());
+    return query;
+  },
 };
 
 module.exports = {
-  $reportName,
+  getAllByFiscal: async (fiscal) => {
+    const [date, report] = await Promise.all([getCurrentDate(), queries[name](fiscal)]);
+
+    return {
+      date,
+      report,
+    };
+  },
 };