+++ backend/templates/create-report/template-files/routes/index.js
@@ -1,21 +1,18 @@
-const controller = require("@controllers/reports/$reportName.js");
+const getControllerFrom = require("@controllers/reports/genericReportController");
 // const validators = require("@validators/report");
-const what = "report";
 
-const routes = [
-  {
-    method: "GET",
-    url: `/${what}/projects/$reportName`,
-    onRequest: controller.$reportName,
-    handler: controller.getReport,
-  },
-];
-const registerRoutes = (fastify, options, done) => {
-  // Ensure all of the routes above get registered.
-  routes.forEach((route) => fastify.route(route));
-  done();
-};
+const name = `$reportName`;
+const controller = getControllerFrom(name);
 
 module.exports = {
-  registerRoutes,
+  registerRoutes: (fastify, options, done) => {
+    fastify.route({
+      method: "GET",
+      url: `/report/projects/${name}`,
+      // schema: getAll,
+      onRequest: controller[name],
+      handler: controller.getReport,
+    });
+    done();
+  },
 };